build: possibly fix the "make -j" wreckage for parsers
The two Bison parsers (dot and rat) overlap on some files, and
therefore must not be run concurrently.  This was addressed by a
dependency between of the two of them.  Unfortunately there is a
spello in these dependencies.

Since I have never faced the problem (possibly because I used
out-of-tree builds only), I can't tell if this really fixes the
problem.

* lib/vcsn/local.mk: Fix spello.
Formatting changes.