efsm: also set the osymbols list for acceptors
Some OpenFST tools seem to really require an output-symbol list, even
for acceptors.  While fstrmepsilon perfectly works with just the
isymbols, fstintersect (equivalent to our vcsn-product: the Hadamard
product) for instance, seems to require the osymbols; this seems to be
due to the fact that Open FST bases its implementation of intersect on
its (transducer) composition.

* bin/efstdecompile, vcsn/algos/efsm.hh: Use --osymbols and --keep_osymbols.
* tests/tafkit/efsm.dir/a1.efsm, tests/tafkit/efsm.dir/binary.efsm: Regen.