scanner: factor the handling of yyalloc, yyrealloc and yyfree
We used to use "%option prefix" to embed two different Flex scanners
in a single program/library.  Now that we use namespaces, this is
useless.

* lib/vcsn/dot/driver.hh, lib/vcsn/dot/driver.cc, lib/vcsn/dot/fwd.hh,
* lib/vcsn/dot/scan.hh, lib/vcsn/dot/scan.ll: No longer used %option,
so simplify names (detail_dotFlexLexer -> yyFlexLexer).

* lib/vcsn/rat/driver.hh, lib/vcsn/rat/driver.cc, lib/vcsn/rat/fwd.hh,
* lib/vcsn/rat/scan.hh, lib/vcsn/rat/scan.ll: Likewise.

Flex declares yyalloc, yyrealloc and yyfree in the output file
(scan.cc) right before the user sections, therefore right before our
YY_FLEX_NAMESPACE_BEGIN, but it implements them within this namespace.
We used to solve this issue in both scanners.  Factor this.

* build-aux/bin/flex++.in: Disable the declarations of yyalloc, yyrealloc
and yyfree.
* lib/vcsn/dot/scan.ll, lib/vcsn/rat/scan.ll: Remove local workarounds,
now that these declarations are disabled.
* vcsn/misc/flex-lexer.hh: Declare them.