weights: provide a means to "promote" a Boolean to another type
This will be used in forthcoming patches that intented to provide
support for product of automata of different weightsets.

* vcsn/core/rat/ratexpset.hh, vcsn/core/rat/ratexpset.hxx,
* vcsn/weights/b.hh, vcsn/weights/q.hh, vcsn/weights/r.hh,
* vcsn/weights/z.hh, vcsn/weights/zmin.hh
(self_type): New, the type of *this.
(conv): New overload, to provide conversion _from_ Boolean (b) to
self_type.
(get_union, intersection): Add support for Boolean as one of the "other"
type.