dyn: start factoring the computation of names
Currently, the handling of signatures of algorithms is specific for
each case.  Try to be more systematic, so that eventually, we have
something easier to factor.

* vcsn/misc/name.hh: New file.
(vnamer, snamer, sname, vname): New.
Generalized version of what used to be in...
* lib/vcsn/algos/registry.hh: here, providing a means to support "plain"
types such as bool.
* vcsn/algos/determinize.hh, lib/vcsn/algos/determinize.cc: Declare the
two arguments of determinize (including the Boolean for completion).
* vcsn/ctx/lal_char_b.hh: Adjust: determinize has now two parameters.
* vcsn/local.mk: Adjust.