minimize: simplify the code
Because we store partitions as vectors of states, it is better to use
a map from classes to vectors of states instead of a multimap from
classes to states, as the latter requires a conversion to a vector of
states.

This goes from about 20s to about 13s on erebus in "intricate1200"
as in tests/score.

* vcsn/algos/minimize.hh (target_class_to_states_t): Now of a map
of set_t.
Adjust dependencies.
(split_class): Remove.