tests: write some of the tests on top of python
This is an experiment: rewrite parts of the tafkit/derivation.chk test
in Python.  Writing the tests is quite comfortable, but the main
benefit is that, of course, running the test case of *much* faster.
Of course: there's no I/O, nor repeating loading/unloading of a
TAF-Kit programs.

* tests/python/derivation.py: New.
* tests/python/local.mk: New.
* tests/local.mk: Adjust.
(BUILDCHECK_ENVIRONMENT): Pass all the needed envvar tweaks.
See also tests/bin/vcsn.in.