build: no longer prefer -stdlib=c++11 over the "standard" C++ library
There's a use storm ahead of us: there are two C++ libraries, Gcc's
libstdc++ and Clang's libc++, and only one of the them is the "system"
library on a given distro.  Which means that C++ libraries such as
Boost are compiled with one of the two, and since they are
incompatible (at the object level), we must _not_ switch.

Currently Mac OS X has a severe problem with this issue (see
https://trac.macports.org/ticket/41588 for instance).  But
"fortunately" on Ubuntu (such as on our build farm), one _can_ use
clang with libstdc++.  So at least to allow using Boost with Clang
with C++11 on Ubuntu, do not "prefer" libc++ over libstdc++.

* build-aux/m4/ax_check_compile_flag.m4: Accept an additional argument
for the source to compile.
* configure.ac: First try without -stdlib, then with.