minimize: import the latest version of (regular) Moore minimization
We try to offer several implementation of the minimization.  Before
the introduction of the signature-based minimization, there was a more
traditional implementation of Moore.  Import it back, and provide a
"top-level" minimize function which accepts the name of the algorithm
to run: "moore" or "signature".

* vcsn/algos/minimize.hh: New, based on the version of commit
e36cae523e62cd5dda8230f4fb2b17dc6b83136d, i.e., not signature based.
(dyn::detail::minimize): Point to both versions.
* bin/vcsn-minimize.cc, lib/vcsn/algos/minimize.cc: Adjust.
* tests/tafkit/minimize.chk: Test both versions.
* vcsn/ctx/lal_char_b.hh, vcsn/dyn/algos.hh, vcsn/local.mk: Adjust.