tests: use the Python interface to tests the ratexp
This is way faster (from 40s to 5s in -j1, from 20s to 5s in -j4), as
we launch the python interpreter only once (per *.rat file) instead of
launching one vcsn-cat per test case.

It turns out that "make check" when the test suite was run, and
therefore there is nothing to run, lasts 5s on my machine.  In other
words, the cost of the tests themselves is completely hidden by the
cost of the Makefile.

* tests/rat/rat: Remove this Perl program, replaced by...
* tests/bin/rat.py: this Python guy.
* tests/rat/local.mk: Adjust.