minimize: moore: recycle class identifiers
Currently when splitting a class, we create new class numbers, which
results in a sparse set of class ids, which is bad if we want to use
class numbers as indices in vectors.  Which we want.

* vcsn/algos/minimize.hh (make_class): Accept the request class number
as optional argument.
(build_classes_): No longer use a temporary list of classes to build,
build them directly.