eval: fix serious bug
Under Python, this return 2 instead of 6:

vcsn.context('lal_char(abc)_z').ratexp('<2>a+<4>a').standard().eval('a')

this is because the automaton build by standard has a hole in its
state numbers, and therefore the array, which is index by state
numbers, but of size num_states() lacks some of the states.

In other words, *never used num_all_states* etc., they are wrong.

This cannot be seen in TAF-Kit, since the "eval" steps parses the
automaton, which removes the holes in the state numbers.

Note using a vector seems stupid here: it would be better to have a
worklist of states being visited.  It will be faster.

* vcsn/algos/eval.hh: Fix the size of the array.