minimize: signature: fix non-determinism
* vcsn/algos/minimize-signature.hh: Add tons of debugging traces, and
fix one true bug: the comparison of two sets of signatures did not
ensure that both sets had the same size.  As a consequence, the
equality comparison was performed only up to the length of the first
set.  It turns out that we have an empty set which appears, and if it
"of course" always equal to any other set.

For some reason libc++ seems to be not as good as libstdc++ is, and
performs many more hash computations, and some useless equality
comparison.  Because of the aforementioned issue, in some cases we
used to consider as equal sets that should have be distinguished.