ratexp: use std::dynamic_pointer_cast more consistently
My previous experience at Gostai with Urbi has shown me that
dynamic_cast is slow, and it's faster to use a combination of a
function that returns the type (which is what type() does here), and
then a static_cast.

However, this is not pleasant to read, and, contrary to Urbi, here,
this is not performance critical.

It is far more readable as follows.  Note however that if we do want
more performances, we can introduce a parameterized function that
first call type(), and then the static_pointer_cast.

* vcsn/core/rat/ratexpset.hxx: Prefer std::dynamic_pointer_cast to
cascades of type() && down_pointer_cast.