labelset: add hashing and comparing classes; weightset: add hashing
Some labelsets are nontrivial to compare and hash, but it's useful to
have them as keys of std::unordered_set, std::unordered_map and
friends; all these template classes take hashing and comparison
classes as optional parameters.  Add such classes, templated on
labelset.  Add a hash static method to each labelset, to be called
from the hasher class.  Add a hash static method to each weightset.

The ratexp case is the most complicated one: hashing depends on
substructures, and is implemented as a visitor; hashing on some
ratexpset requires hasing on its labelset and weightset.

* vcsn/labelset/equalto-labelset.hh: New.
* vcsn/labelset/hasher-labelset.hh: New.

* vcsn/labelset/letterset.hh: Here.
* vcsn/labelset/nullableset.hh: Here.
* vcsn/labelset/oneset.hh: Here.
* vcsn/labelset/wordset.hh: Here.

* vcsn/weights/b.hh: Here.
* vcsn/weights/f2.hh: Here.
* vcsn/weights/polynomialset.hh: Here.
* vcsn/weights/q.hh: Here.
* vcsn/weights/r.hh: Here.
* vcsn/weights/z.hh: Here.
* vcsn/weights/zmin.hh: Here.

* vcsn/core/rat/ratexpset.hh: Here.
* vcsn/core/rat/ratexpset.hxx: Here.

* vcsn/core/rat/hash.hh: Implement hash visitor.
* vcsn/core/rat/hash.hxx: Implement hash visitor.

* vcsn/local.mk: Here.