python: add vcsn_datadir variable, to load lib automatons in python
So far, the automatons installed with the base libraries can be loaded
with tafkit thanks to the use of the environment variable
VCSN_DATADIR. To facilitate the loading of these files in python, a new
variable, vcsn_datadir, has been introduced. The file python/vcsn.py is
now to be configured with the content of the variable.

* configure.ac: Add vcsn.py to generated files
* python/vcsn.py: Rename to python/vcsn.in.py, add vcsn_datadir variable
* tests/bin/vcsn.in: Look for python/vcsn.py in the builddir, not srcdir