tuple hashing: factor it and make it interoperable with misc/hash
Add a hash method in tupleset, and change the equal method to be
static, for compatibility with the hashing facilities in
vcsn/misc/hash.hh.  Add a std::hash specialization for std::tuples.
Move the "seq" hack from tupleset to the new file vcsn/misc/tuple.hh,
since the same definitions are useful in the std::hash specialization
as well.

* vcsn/labelset/tupleset.hh: Here.

* vcsn/misc/tuple.hh: New.

* vcsn/local.mk: Here.