tupleset: allow components to be tupleset too
It is straightforward except for the fact that tupleset is in
vcsn::ctx, and since print_set in defined as a "member friend", it is
defined in vcsn::ctx too, but it invokes vcsn::print_set, where the
others are.  So we need to bounce to it from
vcsn::print_set(tupleset).  Then there are issues about definition
order to make sure that vcsn::ctx::print_set can invoke itself.

It is unclear whether print_set should be a member: it needs internal
details, so it should, on the other hand, it is more pleasant to see
it as a separate algorithm (a bit like what we do with visitors), in
which case it would better to move this definition of print_set for
tupleset in print.hh.  But there will be code duplication...

However, the current situation, where print.hh loads all the headers
of labelsets and weightsets, is also a PITA: too many compile-time
dependencies.  So on second thought, I prefer that we split print_set
from one place to all the corresponding headers of valuesets.

* vcsn/labelset/tupleset.hh (print_set): Declare early.
Make it a genuine member: it makes it easier to ensure proper template
instantiation when tupleset needs tupleset (or at least, that's the only
way I could please clang).
* tests/python/context.py: Compile tuplesets of tuplesets.