add weighted minimization as a minimization variant for all lal contexts
The implementation relies on the same signature idea employed in
minimize-signature, but in this case the bitmap optimization doesn't
apply, and we really have to store a class-to-weight map for each
output character of each state.  The new "weighted" minimization is
the default variant for all non-Boolean contexts, and from Python;
"signature" remains available for all Boolean automata, and "moore"
for deterministic Boolean automata.  Add tests.

* vcsn/algos/minimize-weighted.hh: New.

* vcsn/algos/minimize.hh: Check weightset at algorithm selection time.
* vcsn/algos/minimize-signature.hh: Don't fail on non-boolean weights.

* vcsn/algos/quotient.hh: Use weights when building result.

* vcsn/ctx/lal_char_b.hh: Remove registration from lal_char_b...
* vcsn/ctx/lal_char.hh: ...Make minimization available in all lal contexts.

* vcsn/local.mk: Here.

* python/vcsn_python.cc: Make 'weighted' the default variant from Python.

* tests/python/minimize.py: Here.
* tests/python/minimize.dir/small-weighted.exp.gv: New.
* tests/python/minimize.dir/small-weighted.gv: New.
* tests/python/local.mk: Here.

* NEWS.txt: Here.