derived-term: use a hash instead of a map
There is no measurable impact at all on erebus with Clang 3.4 (there
might be one with GCC, whose hashes seemed to behave better (fewer
calls to the hash function)).  This is unexpected, yet understandable:
to hash does need to traverse the ratexp, and then a second traversal
for "equals".  It also shows that most of the poor performances for
derived_term are rather in derivation.

* vcsn/algos/derivation.hh (derived_term): Use a hash.