polynomialset: parse and display brackets in transitions
For the labels that use an alphabet (lan, lal and law), it is now possible to
write [a-z] which means "every letter between a and z in the alphabet".
The displaying of labels in transitions is also simplified: every transition is
always represented in the way that takes the lesser space ("a, b" but "[a-c]").

* vcsn/labelset/genset-labelset.hh: Add convs_() to parse brackets
* vcsn/labelset/letterset.hh: Add convs() for letters
* vcsn/labelset/nullableset.hh: Add convs() for nullables
* vcsn/labelset/wordset.hh: Add convs() for words
* vcsn/labelset/oneset.hh: Add convs() not-implemented empty function
* vcsn/labelset/tupleset.hh: Idem
* vcsn/weights/polynomialset.hh: Here
* tests/tafkit/derivation.dir/h3-dt.gv: Fix tests to match with the new display
* tests/tafkit/de-bruijn.chk: Idem
* tests/tafkit/determinize.dir/ladybird-4-det.gv: Idem
* tests/tafkit/determinize.dir/ladybird-8-det.gv: Idem
* tests/tafkit/sum.chk: Idem
* tests/tafkit/universal.chk: Idem