chain: new algorithm on ratexp
I'd like to call it power, but its semantics has nothing to do with
that of automata.  There, it is called chain.  Basically, our problem
is that "mul" seems to denote intersection for automaton, not
concatenation.  IMHO, it should rather be concatenation, and we should
use "intersection", or maybe "and_" for the Hadamard product.

* lib/vcsn/algos/concatenate.cc, python/vcsn_python.cc,
* vcsn/algos/concatenate.hh, vcsn/ctx/instantiate.hh,
* vcsn/dyn/algos.hh (chain): New algorithm.

Conflicts:
	python/vcsn_python.cc