ratexp: shuffle: add trivial identities about \e
Having a neutral is helpful in some traversals: instead of having to
start with the first element, and then work on the tail, we can start
by the neutral, and then work on the whole list.

For an example of this, see the (forthcoming) commit introducing
"first-order", an alternative to "derivation".

* vcsn/core/rat/ratexpset.hxx: here.