labelsets: move them out of ctx::
The labelsets, and the contexts, are truly part of Vaucanson, and I
now adopt the opinion of Alexandre Duret-Lutz: let's keep these guys
in vcsn::.  Besides the current status was really wrong, with
weightsets in vcsn::, and labelsets in vcsn::ctx::.  This is really
wrong for tupleset for instance.

context should be moved too.

* vcsn/labelset/genset-labelset.hh: Move from ctx:: to detail::.
* vcsn/labelset/letterset.hh, vcsn/labelset/nullableset.hh,
* vcsn/labelset/wordset.hh: Move out of ctx::.

* tests/unit/label.cc, tests/unit/transpose.cc, vcsn/algos/enumerate.hh,
* vcsn/dyn/translate.hh, vcsn/factory/random.hh, vcsn/labelset/fwd.hh:
Adjust.