alphabet: change special letter from 255 to 127
Depending on the interpretation of char (signed or unsigned), the special
letter was 255 or -1, and the -1 would cause it to be on the first
sorted transitions. For some algorithms, however, we need the one
transitions to be first. Moreover, 127 is DEL and will never be typed by
a user, whereas 255 is y with a diaeresis, so in the case of linguistic
studies, it could be used.

* vcsn/aphabets/char.hh