print_set: make them genuine member functions
I much prefer free standing functions when I can, however when
templated they can create problems when they call functions that are
defined after themselves.  This is for instance the case of print_set
for contexts that needs print_set to be defined for the labelset and
weightset used by the context.

This is too much of a pain: make print_set a member function.

* vcsn/algos/print.hh, vcsn/alphabets/setalpha.hh,
* vcsn/ctx/context.hh, vcsn/ctx/lao_z.hh, vcsn/labelset/letterset.hh,
* vcsn/labelset/nullableset.hh, vcsn/labelset/oneset.hh,
* vcsn/labelset/tupleset.hh, vcsn/labelset/wordset.hh,
* vcsn/weights/b.hh, vcsn/weights/f2.hh, vcsn/weights/q.hh,
* vcsn/weights/r.hh, vcsn/weights/z.hh, vcsn/weights/zmin.hh,
* vcsn/core/rat/ratexpset.hh:
Do it.