tests: avoid race conditions during distcheck
The BF often tests several distchecks for several versions of
Vaucanson in parallel.  The distcheck runs to installcheck target, in
which case the plugin directory is ~/.vcsn/plugins.  So it does happen
that several different versions of Vaucanson all work concurrently in
~/.vcsn/plugins, generate their code, and dlopen... some random
version of the plugins.

I first considered that it is distcheck that should not work in
~/.vcsn, but on second thought, installcheck should not.  And it is
actually much easier with Automake to control installcheck than
distcheck.

* lib/vcsn/dyn/translate.cc: Fix the handling of VCSN_HOME.
* tests/local.mk: Define it for installcheck.