product: use a state map instead of a state multimap
Some experimentations (where the key are pairs of integers, which is
the case for binary product) revealed a disturbing fact: in same case,
trees are significantly faster than hashes.  So use a map instead of
an unordered_map.  We get a minor improvement.

master (v2b.3-261-gb2ca65b)
       4     5     6     7
 4  0.02  0.09  0.30  1.10
 5  0.08  0.29  1.08  4.23
 6  0.29  1.21  4.16 16.30
 7  1.12  4.33 16.46 92.45

With unordered_map

       4     5     6     7
 4  0.03  0.11  0.41  1.43
 5  0.10  0.47  1.53  5.44
 6  0.38  1.38  5.30 21.00
 7  1.42  5.35 21.07 109.95

With map:
       4     5     6     7
 4  0.03  0.10  0.35  1.30
 5  0.10  0.33  1.27  5.21
 6  0.33  1.28  5.17 19.16
 7  1.26  4.98 19.28 100.70

* vcsn/algos/product.hh (transition_map::maps_): Now a map.
Use Meyers' technique for efficient insertion/retrieval.