weights: introduce a (factored) variadic mul
All the weightsets define a binary mul() and a nullary one().  On top
of them it is desirable to offer a variadic mul().  In the
experimentation below, it is in the course of state elimination that
we produce a ratexp "e.f*.g", which is more pleasant as a variadic
product rather than two binary products.  Note, however, that there is
a real need for variadic mul() when implementing a variadic product of
automata.

There are several options to provide such a (factored) mul().

(macro) Define a macro that defines it, and use that macro in the all
the weightsets.  This is unpleasant for obvious reasons, such as
debugging.

(crtp) Use the curiously recurring template pattern: define a "struct
WeightSet<Derived>: Derived" class template, and have all the
weightsets derive from it.  This is nice, but there is one problem:
the value_t of the derived class (e.g., q and its value_t which is an
inner class that defines fractions) must be defined outside of q.  So
it can no longer be an inner class, and traits must be used in order
for WeightSet<Derived> to know Derived::value_t.

(mixin) Define a "variadic_mul_mixin<WeightSet>: WeightSet" class
template that adds features to the "base" WeightSet.  This avoids the
need for traits and allows value_t to remain an inner class.  However,
it also means that q is no longer the right type to use, rather "q :=
variadic_mul_mixin<q>".  So rename the latter as "q_impl" to avoid bad
surprises and unreadable error messages (I mean, even if we move
q_impl into detail::, it would probably be a bad idea to name it
detail::q).

This implements (mixin).  There is one big caveat: "q_impl::self_type"
must be defined a "q", not "q_impl".  Or, more precisely,
"q_impl::make()" must return a "q", not a "q_impl".

* vcsn/weightset/weightset.hh: New.

* vcsn/weightset/b.hh, vcsn/weightset/f2.hh,
* vcsn/weightset/q.hh, vcsn/weightset/r.hh,
* vcsn/weightset/z.hh, vcsn/weightset/zmin.hh (b, f2, q, r, z, zmin): Rename as...
(b_impl, f2_impl, q_impl, r_impl, z_impl, zmin_impl): these.
* vcsn/weightset/fwd.hh: Adjust.
(b, f2, q, r, z, zmin): Define them.

* vcsn/core/rat/ratexpset.hh, vcsn/core/rat/ratexpset.hxx (ratexpset): Rename as...
(ratexpset_impl): this.
* vcsn/core/rat/fwd.hh: Adjust.
(ratexpset): Define it.

* vcsn/algos/aut-to-exp.hh: Use variadic ws.mul.