build: let's put GCC back in business
There is a workaround for
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=51253:

This fails:
  struct swallow
  {
    template<typename... Types>
    swallow(Types &&...){}
  };

  template<int... IS>
  int f1()
  {
    int i = 0;
    swallow{ i = 10 * i + IS... };
    return i;
  }

This passes:

template<int... IS>
int f2()
{
  using swallow = int[];
  int i = 0;
  (void) swallow{ i = 10 * i + IS... };
  return i;
}

It clutters the code a little bit, as we need some final value (here
an int), but it is moderate.  This does not entirely free us from the
original problem which affects routines such as make_tuple, which is
used in our input routines.

* vcsn/algos/product.hh, vcsn/labelset/tupleset.hh,
* vcsn/misc/tuple.hh, vcsn/weightset/weightset.hh: Use this workaround.

* vcsn/misc/zip-maps.hh: Please GCC: add a this->.