eval: use dyn::label
Now that we finally have dyn::label, fix the signature of the routines
that have labels as argument.  Start with eval.  Eventually it should
be possible to apply evaluate to (some) k-band automata.

* vcsn/algos/eval.hh (eval): Take a label as input, instead of a
std::string.
Adjust dependencies.

* python/vcsn_cxx.cc (eval): Replace with...
(eval_): this, that takes a label.
* python/vcsn/automaton.py (automaton.eval): New, performs the
conversion from the string (or whatever) to label.

* bin/parse-args.hh, bin/parse-args.cc (read_label): New.
* bin/vcsn-tafkit.cc (evaluate): Use it.

* tests/demo/prod-eval.cc (read_label): New.