labelset: let genset() return the genset object, not a shared_ptr
Things are still somewhat unclear to me wrt labelset() and
weightset(): they return pointers.  In the case of a labelset's
genset(), it seems more natural to be able to iterate directly on it.
However the real motivation here is that the genset() of a tupleset
does not need to be a pointer (actually it is harder to make it a
pointer).  So for consistency, return it as a genuine value.

* vcsn/labelset/genset-labelset.hh, vcsn/labelset/letterset.hh,
* vcsn/labelset/nullableset.hh, vcsn/labelset/wordset.hh (genset): Return
a genset_t, not a genset_ptr.
Adjust dependencies.