compose: new composition algorithm
It is now possible to compute the composition of two transducers. However, the
only possibility is to take the second band of the left transducer and the
first band of the second transducer. Moreover, the result will be invalid most
of the time if the concerned band has epsilon-transitions (though thompson is
fine).

* lib/vcsn/algos/compose.cc, vcsn/algos/compose.hh: New.
* lib/vcsn/local.mk, vcsn/local.mk, vcsn/ctx/instantiate.hh, vcsn/dyn/algos.hh,
* python/vcsn_cxx.cc: Register.
* tests/python/compose.hh: New.