polynomialset: make it usable as a weightset
A user can now use polynomialsets as weightsets.  The syntax is
weightset ::= polynomialset<context>

The context parameter requires a labelset with a unit and a product,
so at this time polynomials are only instantiatable on contexts whose
labelset is lao, law, ratexpset, or a lat whose components are all
suitable.

* vcsn/weightset/polynomialset.hh: Here.
* vcsn/weightset/fwd.hh: Here.

* vcsn/dyn/context-parser.hh: Here.
* lib/vcsn/dyn/context-parser.cc: Here.

* NEWS.txt: Here.