random: add a random_uniforn algorithm
The random_uniform(N) algorithm generates a complete automaton with N states,
and creates a transition from all states for all labels in the alphabet to a
random state, chosen by uniform distribution.

Rationale: generating an automaton with this algorithm gives a very high
probability of having a synchronizing automaton[1]. This will be very useful
for test and research purposes.

[1] On the probability of being synchronizable, Mikhail V. Berlinkov:
    http://arxiv.org/abs/1304.5774

* vcsn/factory/random.hh: Core algorithm.
* vcsn/dyn/algos.hh: Declare it.
* lib/vcsn/algos/random.cc: Call it from registry.
* vcsn/ctx/lal_char.hh: Instantiate it.
* python/vcsn_cxx.cc: Python binding.