labelset: add "letters_of", a means to iterate over letters of a word
Suggested by Alexandre Duret-Lutz.  This is a much more consistent
approach that just expecting the word_t to be iterable.  This is
especially visible with word_t of 'lat<lal_char(ab), lal_char(xy)>':
a std::tuple("ab", "xy") is not directly iterable as
{std::tuple('a', 'x'), std::tuple('b', 'y')}.

Another possibility would have been to introduce iterable_tuples and
use these instead of std::tuple as word_t, but this proposal is much
simpler to implement.  It is also somewhat less nice of the user code
side, as it was nice to simply write

      for (auto l: word)

but really, it is more consistent with Value/ValueSet to expose this
iteration:

     for (auto l: ls.letters_of(word))

* vcsn/labelset/letterset.hh, vcsn/labelset/nullableset.hh,
* vcsn/labelset/wordset.hh, vcsn/labelset/tupleset.hh (letters_of): New.

* vcsn/algos/synchronizing-word.hh: Use it.