algos: avoid "static" variables like hell
Outputting was running invalid code (diagnosed by Asan): the problem
was the alphabet was abcde, but the labels used were abcdefghijk.  It
turns out that previously a context with the short alphabet was
defined, used for printing, and then another context (of same type)
had a wider alphabet, but because of "static", the _previous_
polynomialset (with the old, short, alphabet) was used!

* vcsn/algos/efsm.hh, vcsn/algos/first-order.hh, vcsn/algos/grail.hh:
Don't use static variables.