polynomialset: simplify the printing with ranges
* vcsn/weightset/polynomialset.hh: There is no need to sort the
letters, since the polynomial has its labels sorted.
Likewise, the constant-term is unique and first: there is no reason
to loop on the whole polynomial.
(print_with_ranges_): Handle the constant-term here instead of leaving
it to the callers.
This allows to avoid having to duplicating a LAN polynomial to extract
its LAL counterpart.
Remove the useless enable_if, and corresponding templatization.
(print_): Can be factored bw LAN and LAL.