kmul: generalize to any automaton, and treat 0 explicitly
The right-multiplication by a weight uselessly required a standard
automaton; it now accepts any automaton.  The left-multiplication
keeps its specification for standard automata, but now supports
non-standard automata, in which case the weight is put on the initial
arrows.

The multiplication of an automaton by the null weight results in the
"zero automaton": a single state, initial, and no transitions.

* vcsn/algos/left-mult.hh: Do that.
* tests/python/test.py (rst_diff): New, extracted from...
(CHECK_EQ): here.
(CHECK_ISOMORPHIC): New.
* tests/python/kmul.py: Refactor old tests to be more readable.