expansionset: extract from first-order
The current implementation of first-order is somewhat hard to read,
although some parts should be really straightforward to read.  This is
because "expansions" (my current name to denote the core data
structure) are dealt with inside first_order_visitor.  Sure, some
routines are already decoupling the visitor from the handling of these
expansions, but that's not enough.  So let's start writing this
expansion support, up to Python.

* vcsn/algos/first-order.hh (expansionset): New.
(first_order_visitor): Move some parts into...
(expansionset): here.