zip-maps: work on const structures
I need to be able to browse const-structures, and arguably, it's
probably the only choice that makes sense.  However, it would be
better to design these auxiliary iteration tools so that they would
also support non-const iteration.  However, I don't have time for this
now, and then again, it's not clear we would need the feature anyway.

* vcsn/misc/zip-maps.hh: Propose only a const_iterator interface.