rat: revamp character-classes
Instead of handling character classes in the scanner, do it in the
parser.  This is more elegant, but also allows to deal with character
classes that accept single-quoted letters, as is the case for instance
with tuples of labelsets.

While at it, accept `[]' to denote 'any letter of the alphabet'.

* vcsn/core/rat/ratexpset.hh, vcsn/core/rat/ratexpset.hxx,
* vcsn/dyn/ratexpset.hh, vcsn/dyn/ratexpset.hxx (char_class): New.
* lib/vcsn/rat/scan.ll: No longer handle the character classes in
the scanner, but...
* lib/vcsn/rat/parse.yy: here, in the parser.

* tests/rat/b.rat: Check.