first-order: fix the expansion of transposition of variadic product
This has long been bothering me: "d(abcd) = a . (bcd)".  Nice and
beautiful.  Unfortunately we used to have "d^T(abcd) = d . (c(ab){T})".

The reason was that even in the first case, a . (bcd), the polynomial
part (bcd) was built in two steps: first b then the ending range, cd,
and both were concatenated together into a single variadic product:
bcd.  In the transposed case, this last step did not work: we get c{T}
= c, then the transposed ending range (ab){T}, concatenated together
into c(ab){T}.

Actually the case of the null resulting weight should be treated
earlier in the loop body, so we directly build the range bcd in the
first case, and the transposed range (abc){T} in the second case.

The code is even better looking this way.

* vcsn/algos/first-order.hh (prod): Be more cautious when
short-circuiting the remained of the loop.

* tests/python/first-order.py: Adjust expectations, finally!