dyn: fix the argument order of print for labels
I have long been ambivalent on print: should it be print(obj, stream),
or print(stream, obj)?  The latter is quite natural to read, and the
former is more consistant with the value/valueset scheme where all (or
almost) the function start by the value.

So I have tried both orders to see.  There are two additional
arguments: it makes sense to have a default stream (std::cout), which
is also the case in Python, where the value is the first value.

So let's carve this in stone: the interface for the print functions is
to have the value first, then the stream.

* lib/vcsn/algos/print.cc, python/vcsn_cxx.cc, vcsn/algos/print.hh,
* vcsn/dyn/algos.hh: Fix the order of print for labels.