rat: symmetrically to lweight, associate rweight to a single letter
So far we used to parse "ab<x>" as "(ab)<x>".  Now it is parsed as
"a(b<x>)", so thanks to trivial identities, it will become "a(<x>b)"
which is printed as "a<x>b".  This is again more natural wrt
derivation:

   q.ratexp('(ab)<3/4>').derived_term().ratexp() => a<3/4>b

It's also easier to remember to have all the postfix operators behave
the same way.

* lib/vcsn/rat/parse.yy: Give "weight" a higher precedence than CONCAT.
* vcsn/core/rat/ratexpset.hxx (concat_): Be both more general and simpler.

* tests/python/aut-to-exp.py, tests/python/first-order.py,
* tests/python/kmul.py, tests/python/standard.py,
* tests/rat/br.rat, tests/rat/z.rat: Adjust expectations and/or
inputs.