first-order: disable incorrect code for ldiv
I have no idea how I came to believe that (e*)\(f*) = (e\f)*f*, but
it's dead wrong, as demonstrated by a*\(ab)* for instance, which is
certainly not the same as (a\ab)*(ab)* = b*(ab)* (which accepts bb for
instance).

* vcsn/algos/first-order.hh: Disable this rubish.
This approach of ldiv (without \e) is to be removed completely, obviously,
as it fails to terminate properly.  But there are still a couple of
possibilities I want to explore before full removal.
* tests/python/first-order.py: Disable tests where we fail to terminate
(or rather, where we terminate thanks to SIGSTACK).