todo: remove dead entry about precedence
See the commits f8bfeec and 432d60e.  We have to have two different
parses of 'ab<3>': in LAL it was "(ab)<3>" ('(ab)' is a product), and
"<3>(ab)" in LAW ('(ab)' is a word).

   $ vcsn-cat -C 'lal_char(ab)_z' -Ee 'ab<3>'
   a<3>b
   $ vcsn-cat -C 'law_char(ab)_z' -Ee 'ab<3>'
   <3>(ab)

But now, since the changes in precedences, "ab<3>" is consistently
read "a<3>b" (which means "a(<3>b)" in both cases, and, of course,
"(ab)<3>" is also read the same way in both cases (though one with a
word, the other with a product).

* TODO.txt: Remove dead issue.