pair: provide a means to keep the initial states
Add a keep_initials option to the pair algorithm, to keep the states from
the original automaton.
In the new transition function d' of the pair automaton generated with this
option enabled, if d(p, a) == d(q, a) then d'({p, q}, a) = d(p, a) instead of
q0.

* vcsn/algos/synchronizing-word.hh: Here
* lib/vcsn/algos/synchronizing-word.cc: Change signature
* vcsn/ctx/instantiate.hh: Change signature
* vcsn/dyn/algos.hh: Change signature
* python/vcsn_cxx.cc: Change signature, add overload