automaton: handle them via shared pointer
This is a massive commit with way too many changes for me to spell
them down completely.

The change is basically hiding automata structures (e.g.,
mutable_automaton, automaton_decorator, transpose_automaton, etc.)
into detail:: and with _impl appended (to avoid issues with algorithms
implemented in detail, where it would be painful to have to spell out
vcsn::detail::mutable_automaton as opposed to vcsn::mutable_automaton:
it is safer and clearer to have vcsn::mutable_automaton and
vcsn::detail::mutable_automaton_impl).  Then a typedef restores the
previous names as shared_ptr to the corresponding structures.

Then, most of the changes consist in adjustement: add the needed
std::make_shared where needed, and many many many aut.foo() that
become aut->foo().

The core motivation for this change is that memory management with the
wrappers (transpose_automaton, blind_automaton, and others are to
come) cannot be done properly.  Currently a wrapper _hopes_ that the
wrappee will not be destroyed with the wrapper uses it.  We also want
to get back the wrappee (when transposing a second time for instance,
etc.).  Then again, we have an issue with memory management.

Hence the move to pointers.

I observe no significant changes in the benches on erebus, Clang 3.5
-O3 -DNDEBUG (rhs is after this commit, but of course the SHA1 is now
wrong):

$ vcsn score-compare +scores/35s/v2b.3-599-g2270105 +scores/35s/v2b.3-623-g87ec898
4.18s  4.20s  a & 12               # a = std([a-e]*b(<2>[a-e])*)
7.71s  7.68s  a.determinize()      # a = ladybird(21)
0.53s  0.53s  a.eval("a"*151)      # a = de_bruijn(150), 1000x
0.37s  0.36s  a.format("dot")      # a = std(a?{500})
0.13s  0.13s  a.format("efsm")     # a = std(a?{500})
0.11s  0.10s  a.format("fado")     # a = std(a?{500})
0.10s  0.10s  a.format("grail")    # a = std(a?{500})
0.34s  0.33s  a.format("tikz")     # a = std(a?{500})
3.33s  3.24s  a.infiltration(a)    # a = std([a-e]?{30})
2.80s  2.81s  a.minimize("moore")  # a = std([a-k]{2000})
4.39s  4.47s  a.minimize("signature") # a = std([a-g]{800})
3.89s  3.91s  a.product(a)         # a = std([a-e]?{50})
2.35s  2.37s  a.proper()           # a = thompson(a?{1200})
0.72s  0.72s  a.ratexp()           # a = std([a-d]?{100})
2.34s  2.37s  a.shortest(5)        # a = de_bruijn(9)
3.03s  2.99s  a.shuffle(a)         # a = std([a-e]?{50})
1.23s  1.28s  a.sort()             # a = std([a-e]?{600})
1.00s  1.02s  a.synchronizing_word() # a = de_bruijn(6)
0.15s  0.15s  b.format("text")     # b = lal_char(abc)_b, 100000x
0.59s  0.59s  b.ratexp(e)          # e = "(\e+a)" * 500, 100x
3.19s  3.20s  r.derived_term()     # r = (a+b)*b(<2>a+<2>b){150}   on [a-z]  -> Z
2.82s  2.84s  r.derived_term()     # r = (a+b)*b(<2>a+<2>b){150}   on [ab]   -> Z
0.16s  0.16s  r.format("text")     # r = b.ratexp("(\e+a)" * 500), 1000x
0.22s  0.22s  r.linear()           # r = (a+b)*b(<2>a+<2>b){150}   on [a-z]  -> Z
0.04s  0.03s  r.standard()         # r = (a+b)*b(<2>a+<2>b){20000} on [a-z]  -> Z
0.04s  0.04s  r.thompson()         # r = (a+b)*b(<2>a+<2>b){20000} on [a-z]? -> Z
1.06s  1.09s  read(s)              # s = dot(std(a?{500}))
0.23s  0.24s  read(s)              # s = efsm(std(a?{500}))
0.15s  0.16s  read(s)              # s = fado(std(a?{500}))