first-order: introduce the linear-automaton as intermediate
Current first-order computes an automaton maintaining the
correspondance between ratexps and states itself.  Now that we have
automaton-decorators, it is more natural, and more powerful, to
present the interface of an automaton whose states are labeled by
ratexps.  This way the algorithm itself (construction of the
derived-term automaton on top of first-order) no longer maintains
anything related to the automaton, it just builds it using ratexps as
states.

Note that this provides a fine replacement of the "origins" approach:
the metadata are now properly stored in the result, instead of being
sent to stdout...

Better yet: at some point we should be able to invert the control and
have the automaton call the expansion algorithm.  This will become a
very natural implementation of a lazy construction of the derived-term
automaton.

Of course, many other algorithms (actually, almost all the algorithms
that feature origins support) are to be converted to something like
this in the future.

* lib/vcsn/dyn/context-parser.cc, lib/vcsn/dyn/context-printer.cc:
Support linear_automaton.
* vcsn/algos/first-order.hh (linear_automaton): New.
(first_order): Use it
* vcsn/algos/fwd.hh: Adjust.